requirements.txt
easyocr==1.6.2
opencv-python-headless==4.7.0.72
numpy==1.25.2
pillow==10.0.0
streamlit==1.22.0
imutils==0.6.0
pytest==7.4.0
